requirements.txt
flask
flask-cors
requests
beautifulsoup4
openai
shopifyapi